+++ test/web-auth/cross-origin-authentication.test.js
@@ -44,7 +44,7 @@ describe('auth0.WebAuth.crossOriginAuthentication', function() {
         storage.setItem.restore();
       }
     });
-    it('should call /co/authenticate and redirect to /authorize with login_ticket', function() {
+    it('should call /co/authenticate and redirect to /authorize with login_ticket using `username`', function() {
       stub(request, 'post', function(url) {
         expect(url).to.be('https://me.auth0.com/co/authenticate');
         return new RequestMock({
@@ -70,6 +70,41 @@ describe('auth0.WebAuth.crossOriginAuthentication', function() {
       });
       this.co.login({
         username: 'me@example.com',
+        password: '123456',
+        anotherOption: 'foobar'
+      });
+      expect(this.webAuthSpy.authorize.getCall(0).args[0]).to.be.eql({
+        username: 'me@example.com',
+        loginTicket: 'a_login_ticket',
+        anotherOption: 'foobar'
+      });
+    });
+    it('should call /co/authenticate and redirect to /authorize with login_ticket using `email`', function() {
+      stub(request, 'post', function(url) {
+        expect(url).to.be('https://me.auth0.com/co/authenticate');
+        return new RequestMock({
+          body: {
+            client_id: '...',
+            credential_type: 'password',
+            username: 'me@example.com',
+            password: '123456'
+          },
+          headers: {
+            'Content-Type': 'application/json'
+          },
+          cb: function(cb) {
+            cb(null, {
+              body: {
+                login_ticket: 'a_login_ticket',
+                co_verifier: 'co_verifier',
+                co_id: 'co_id'
+              }
+            });
+          }
+        });
+      });
+      this.co.login({
+        email: 'me@example.com',
         password: '123456',
         anotherOption: 'foobar'
       });
@@ -166,7 +201,9 @@ describe('auth0.WebAuth.crossOriginAuthentication', function() {
               error_description: 'a huge error string'
             },
             code: 'any error',
-            description: 'a huge error string'
+            description: 'a huge error string',
+            error: 'any error',
+            error_description: 'a huge error string'
           });
           done();
         }
@@ -320,7 +357,9 @@ describe('auth0.WebAuth.crossOriginAuthentication', function() {
                   error_description: 'a super big error message description'
                 },
                 code: 'any_error',
-                description: 'a super big error message description'
+                description: 'a super big error message description',
+                error: 'any_error',
+                error_description: 'a super big error message description'
               });
               expect(_this.webAuthSpy.authorize.called).to.be.eql(false);
               done();
@@ -359,7 +398,9 @@ describe('auth0.WebAuth.crossOriginAuthentication', function() {
                   error_description: '{"some":"error"}'
                 },
                 code: 'request_error',
-                description: '{"some":"error"}'
+                description: '{"some":"error"}',
+                error: 'request_error',
+                error_description: '{"some":"error"}'
               });
               expect(_this.webAuthSpy.authorize.called).to.be.eql(false);
               done();